requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
attrs==21.2.0
Babel==2.8.0
bandwidth-sdk==12.0.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cleo==0.8.1
click==8.0.0
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
configparser==5.0.2
crashtest==0.3.1
cryptography==3.4.7
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.6
distorm3==3.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.2.16
duplicity==0.8.21
enum34==1.1.10
fasteners==0.14.1
filelock==3.6.0
Flask==2.0.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
ipython_genutils==0.2.0
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.1
jsonpickle==0.9.6
jsonschema==3.2.0
jupyter-core==4.9.1
jwt==1.2.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.6.3
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.2
nbformat==5.1.3
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
OWSLib==0.25.0
packaging==21.3
pandas==1.4.4
paramiko==2.9.3
pastel==0.2.1
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
plotly==5.4.0
ply==3.11
poetry-core==1.0.7
protobuf==3.12.4
psycopg2==2.9.2
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.20
pycryptodome==3.14.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproj==3.3.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
QScintilla==2.11.6
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Shapely==1.8.1.post1
shellingham==1.4.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
sympy==1.9
systemd-python==234
tenacity==6.3.1
tomlkit==0.9.2
traitlets==5.1.1
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.4
usb-creator==0.3.7
virtualenv==20.16.7
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
